fixed bug in hashtables which did not support hashtables having >2^16 entries, added statistics for hashtables
git-svn-id: file:///Users/braun/svn/vermont/branches/vermont-dynamic-config@1580 aef3b71b-58ee-0310-9ba9-8811b9f0742f

+++ concentrator/FlowHashtable.cpp
@@ -295,7 +295,7 @@ int FlowHashtable::aggregateFlow(IpfixRecord::Data* baseFlow, IpfixRecord::Data*
 /**
  * Returns a hash value corresponding to all variable, non-aggregatable fields of a flow
  */
-uint16_t FlowHashtable::getHash(IpfixRecord::Data* data, bool reverse) {
+uint32_t FlowHashtable::getHash(IpfixRecord::Data* data, bool reverse) {
 	int i;
 
 	uint32_t hash = 0;
